Add core library usage example
