+++ lib/src/main/cpp/examples/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(example example.cpp)
+
+target_link_libraries(example sgbm_android_core)
